requirements.txt
marvin
python-dotenv
google-generativeai
groq
flask
markdown
schedule
